Added auth to endpoints, linked user to trip

+++ Api/JourneyHub.Api/Services/TripServices.cs
@@ -1,11 +1,13 @@
 ﻿using AutoMapper;
 using JourneyHub.Api.Services.Interfaces;
+using JourneyHub.Common.Constants;
 using JourneyHub.Common.Models.Domain;
 using JourneyHub.Common.Models.Dtos.Requests;
 using JourneyHub.Data;
 using Microsoft.EntityFrameworkCore;
 using System.Net.Http.Headers;
 using System.Text.RegularExpressions;
+using JourneyHub.Common.Exceptions;
 
 namespace JourneyHub.Api.Services
 {
@@ -20,9 +22,11 @@ namespace JourneyHub.Api.Services
             _mapper = mapper;
         }
 
-        public async Task<Trip> CreateTripAsync(PostTripRequestDto tripDto)
+        public async Task<Trip> CreateTripAsync(PostTripRequestDto tripDto, string userId)
         {
             Trip trip = _mapper.Map<Trip>(tripDto);
+
+            trip.UserId = userId;
 
             trip.Area = await getAreaByCoordinatesAsync(tripDto.MapPoints[0]);
             _context.Trips.Add(trip);
@@ -53,12 +57,18 @@ namespace JourneyHub.Api.Services
             return await _context.Trips.FindAsync(id);
         }
 
-        public async Task<bool> DeleteTripAsync(int id)
+        public async Task<bool> DeleteTripAsync(int id, string userId)
         {
             var trip = await _context.Trips.FindAsync(id);
+
             if (trip == null)
             {
                 return false;
+            }
+
+            if (trip.UserId != userId)
+            {
+                throw new UnauthorizedException(ErrorMessages.Unauthorized_Trip_Deletion, 403);
             }
 
             _context.Trips.Remove(trip);
